tasks-service/requirements.txt
flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.8
requests==2.32.3
tenacity==8.5.0
pybreaker==1.2.0
